refactor: move available functions
